Moved packer type definitions to subdirectory in module, updated README
