+++ packer/types/WavefrontDataPacker.cpp
